Exercises 4.21

+++ blog/controllers/blogs.js
@@ -65,8 +65,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (
+    !request.token ||
+    blog.user.toString() !== decodedToken.toString()
+  ) {
+    return response
+      .status(401)
+      .json({ error: 'token missing or invalid' })
+  } else {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
